requirements.txt
appdirs==1.4.2
Django==1.10.6
django-cors-headers==2.0.2
djangorestframework==3.5.4
gunicorn==19.7.0
packaging==16.8
pyparsing==2.2.0
requests==2.13.0
six==1.10.0
slackclient==1.0.5
websocket-client==0.40.0
